# app/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RustAcademy Backend API
  description: >
    RustAcademy is a Stellar-based payment platform enabling payment links,
    recurring subscriptions, username resolution, marketplace trading,
    fiat on/off ramps, and webhook-driven event notifications.

    **Base URL (production):** `https://api.RustAcademy.example.com`
    **Base URL (development):** `http://localhost:3000`

    Interactive Swagger UI is available at `/docs` when the server is running.
  version: 1.0.0
  contact:
    name: RustAcademy API Support
    url: https://github.com/RustAcademy/RustAcademy
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.RustAcademy.example.com
    description: Production
...
security:
  - ApiKeyAuth: []
  - {}
paths:
...
  /transactions/compose:
    post:
      tags: [transactions]
      operationId: composeTransaction
      summary: Compose a Soroban transaction
      description: Builds and simulates a Soroban contract invocation transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComposeTransactionRequest"
      responses:
        "200":
          description: Composed transaction result
        "400":
          $ref: "#/components/responses/BadRequest"

  # ---------------------------------------------------------------------------
  # Payments
  # ---------------------------------------------------------------------------
...
  /stellar/path-preview:
    post:
      tags: [stellar]
      operationId: pathPreview
      summary: Strict-receive path preview
      description: Returns candidate paths for a fixed destination amount.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathPreviewRequest"
      responses:
        "200":
          description: Path preview results
        "400":
          $ref: "#/components/responses/BadRequest"
  /stellar/path-preview/strict-send:
    post:
      tags: [stellar]
      operationId: strictSendPathPreview
      summary: Strict-send path preview
      description: Returns candidate paths for a fixed source amount.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrictSendPathPreviewRequest"
      responses:
        "200":
          description: Path preview results
        "400":
          $ref: "#/components/responses/BadRequest"
  /stellar/soroban-preflight:
    post:
      tags: [stellar]
      operationId: sorobanPreflight
      summary: Run Soroban preflight simulation
      description: Uses the same pipeline as POST /transactions/compose for health_check.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SorobanPreflightRequest"
      responses:
        "200":
          description: Preflight result
        "503":
          description: RustAcademy_CONTRACT_ID not configured
  /stellar/quote:
    post:
      tags: [stellar]
      operationId: createQuote
      summary: Create a path payment quote
      description: >
        Computes path payment routes with slippage tolerance and TTL.
        Returns a quote ID retrievable until expiry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "200":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /stellar/quote/{quoteId}:
    get:
      tags: [stellar]
      operationId: getQuote
      summary: Retrieve a quote by ID
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          description: Quote expired

  # ---------------------------------------------------------------------------
  # Webhooks
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            request_id:
              type: string
              format: uuid
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            details:
              type: object

    # --- Health ---
...
    ComposeTransactionRequest:
      type: object
      required: [contractId, method, params, sourceAccount]
      properties:
        contractId:
          type: string
        method:
          type: string
        params:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              value: {}
        sourceAccount:
          type: string
        networkPassphrase:
          type: string

    # --- Stellar ---
    AssetRef:
      type: object
      required: [code]
      properties:
        code:
          type: string
          example: USDC
        issuer:
          type: string
          nullable: true
    PathPreviewRequest:
      type: object
      required: [destinationAmount, destinationAsset, sourceAssets]
      properties:
        destinationAmount:
          type: string
          example: "10.5"
        destinationAsset:
          $ref: "#/components/schemas/AssetRef"
        sourceAssets:
          type: array
          items:
            $ref: "#/components/schemas/AssetRef"
          minItems: 1
    StrictSendPathPreviewRequest:
      type: object
      required: [sourceAmount, sourceAsset, destinationAssets]
      properties:
        sourceAmount:
          type: string
          example: "10.5"
        sourceAsset:
          $ref: "#/components/schemas/AssetRef"
        destinationAssets:
          type: array
          items:
            $ref: "#/components/schemas/AssetRef"
          minItems: 1
    SorobanPreflightRequest:
      type: object
      required: [sourceAccount]
      properties:
        sourceAccount:
          type: string
          pattern: "^G[A-Z0-9]{55}$"
    CreateQuoteRequest:
      type: object
      required: [destinationAmount, destinationAsset, sourceAssets]
      properties:
        destinationAmount:
          type: string
        destinationAsset:
          $ref: "#/components/schemas/AssetRef"
        sourceAssets:
          type: array
          items:
            $ref: "#/components/schemas/AssetRef"
        maxSlippageBps:
          type: integer
          minimum: 0
          maximum: 10000
          default: 50
        ttlSeconds:
          type: integer
          minimum: 5
          maximum: 300
          default: 30
        preflight:
          type: boolean
          default: false
    QuotePath:
      type: object
      properties:
        sourceAsset:
          type: string
        sourceAmount:
          type: string
        sourceAmountWithSlippage:
          type: string
        destinationAsset:
          type: string
        destinationAmount:
          type: string
        pathHops:
          type: array
          items:
            type: string
        rateDescription:
          type: string
    QuoteResponse:
      type: object
      properties:
        quoteId:
          type: string
        paths:
          type: array
          items:
            $ref: "#/components/schemas/QuotePath"
        expiresAt:
          type: string
          format: date-time
        maxSlippageBps:
          type: integer
        horizonUrl:
          type: string
        preflight:
          type: object
          properties:
            feasible:
              type: boolean
            error:
              type: string
...